+++ src/window/footer/Footer.tsx
@@ -2,7 +2,7 @@ import { Container, Sprite } from '@inlet/react-pixi'
 import { useContext } from 'react'
 import { footerHeight } from '../globals'
 import { WindowSettingContext } from '../GlobalSetting'
-import barSrc from '../utils/bar.png'
+import barSrc from '../utils/FrameTemplate/bar.png'
 import { Tab } from './Tab'
 
 export const Footer = () => {
