+++ src/window/globals.ts
@@ -3,7 +3,7 @@ export const footerHeight = 54 * scaling
 export const windowHeaderHeight = 32 * scaling
 export const iconSize = 64 * scaling
 export const tabWidth = 200 * scaling
-export const windowEdge = 2 * scaling
+export const windowEdge = 6 * scaling
 
 let volume = 0
 export const setVolume = (v: number) => {
