+++ src/window/utils/FrameTemplate/bar.png
