+++ src/window/utils/FrameTemplate/close.png
