+++ src/window/utils/FrameTemplate/fullscreen.png
