+++ src/window/utils/FrameTemplate/minimize.png
