+++ src/window/utils/MonoColorBG.tsx
